// src/types.ts
export interface VehicleResponse {
  data: {
    name: string;
    type: string;
    message: string;
    status: string;
    detail: {
      registeredPlace: string;
      pucUpTo: string;
      rcStatus: string;
      unladenWt: string;
      hypothecation: boolean;
      financier: string;
      vehicleCategory: string;
      fuelType: string;
      rawFuelType: string;
      registeredAt: string;
      color: string;
      rcNormsDesc: string;
      engineNo: string;
      chassisNo: string;
      chassisNoFull: string;
      insuranceCompany: string;
      insuranceUpTo: string;
      rtoNocIssued: string;
      manufacturingMonthYr: string;
      fitnessUpTo: string;
      taxUpTo: string;
      vehicleClassDesc: string;
      registrationNumber: string;
      modelImageUrl: string;
      seatCap: string;
      insurancePolicyNo: string;
      isCommercial: boolean | null;
      isCommercialFrachiseRegion: boolean;
      brand: {
        make_id: number;
        make_display: string;
      };
      model: {
        model_id: number;
        model_display: string;
      };
      year?: {
        year_id: number;
        year: string;
      };
      regn_year: string;
      states?: {
        state_id: number;
        state_code: string;
        state_name: string;
      };
      RTO?: {
        rto_detail_id: number;
        rto_code: string;
      };
      rc_model: string;
      rc_owner_name: string;
      rc_owner_name_masked: string;
      rc_vh_class_desc: string;
      rc_owner_sr: string;
      makeLogo?: string;
      ds_details?: Array<{
        variant: {
          variant_id: number;
          parent_id: number;
          variant_name: string;
          variant_display_name: string;
          fuel_type: string;
          transmission_type: string;
          transportation_category: string;
        };
        confidence_score: number;
        make_id: number;
        model_id: number;
      }>;
      full_details: {
        color: string;
        maker: string;
        model: string;
        pucNo: string;
        makeId: string;
        source: string;
        grossWt: string;
        modelId: string;
        ownerSr: string;
        pinCode: string;
        pucUpTo: string;
        rtoCode: string;
        seatCap: string;
        taxUpTo: string;
        bodyType: string;
        cubicCap: string;
        engineNo: string;
        fuelType: string;
        mobileNo: string;
        permitNo: string;
        rcStatus: string;
        standCap: string;
        chassisNo: string;
        financier: string;
        unladenWt: string;
        variantId: string;
        wheelbase: string;
        ncrbStatus: string;
        nocDetails: string;
        permitType: string;
        sleeperCap: string;
        blacklisted: string;
        fitnessUpTo: string;
        rcNormsDesc: string;
        rcOwnerName: string;
        noOfCylinder: string;
        rcStatusAsOn: string;
        registeredAt: string;
        rtoNocIssued: string;
        transmission: string;
        vehicleClass: string;
        hypothecation: string;
        insuranceUpTo: string;
        permitIssueDt: string;
        presentAddress: string;
        blacklistStatus: string;
        permitValidFrom: string;
        permitValidUpTo: string;
        rcOwnerLastName: string;
        registeredPlace: string;
        insuranceCompany: string;
        permanentAddress: string;
        rcOwnerFirstName: string;
        vehicleClassDesc: string;
        insurancePolicyNo: string;
        rcOwnerNameMasked: string;
        registrationNumber: string;
        manufacturingMonthYr: string;
        presentAddressMasked: string;
        permanentAddressMasked: string;
      };
    };
    success: boolean;
  };
  meta: any[];
  metadata: {
    timestamp: string;
    user: string;
    user_agent: string;
    total_proxies: number;
    total_user_agents: number;
  };
}